+++ src/tests/mod.rs
@@ -1,6 +1,7 @@
 mod constraint_test;
 mod constraint_test_various;
 mod garde_robotics;
+mod shifted_pose;
 mod test_from_yaml;
 mod test_individual_link_positions;
 mod test_parallelogram;